mtd-sa: Support multiple employments
Previously only a single employment was supported, the employment id was
stored in config.json and used in all API calls that required an
employment id.

Rather than doing that, have the employment id (seid) passed into each
function that requires it, this is then set as params[0] and is picked
out by ep_make_url().

Of course, this is a break in API and thus the library version has been
bumped to 0.1.0 (in pre 1.0 the minor version will be bumped on API
breaks).

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>