mtd: Add defines for the various application connection types
This for part of the fraud prevention[0], there are currently eight
types of application defined and depending on what type you declare
yourself as you need to add various headers.

Currently this just defaults to OTHER_DIRECT and the library ignores
whatever the client/application sets.

Some of the headers would currently be difficult to set under other
connection types, e.g

    Gov-Client-Public-IP

    The public IP address (IPv4 or IPv6) from which the originating
    device makes the request.

doing that is tricky as we wouldn't know that until curl has actually
made a connection...

[0]: https://developer.service.hmrc.gov.uk/api-documentation/docs/fraud-prevention

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>