Start thinking about handling application-restricted endpoints
There are two types of endpoints used in the MTD API

 1) User-restricted
 2) Application-restricted

Currently we are just using user-restricted endpoints, i.e the
self-assessment API.

Lets start thinking about howto handle applications-restricted
endpoints. These are simpler than user-restricted ones as we don't need
to do any user authorisation.

Lets start by adding a function pointer to the curl_ctx structure that
will be set to the appropriate function for handling the refreshing of
the access token for either of the two above endpoint types.

Seeing as most will be of the user-restricted type (I think) lets make
that the default if no function is explicitly set (we will set it to the
current refresh_access_token() in do_curl if ->oauther is NULL).

This can be explicitly set in an endpoint handler function with
something like

    struct curl_ctx ctx = { 0 };

    ...
    ctx.oauther = refresh_application_token();

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>