Change how query string parameters are passed in
Various of the endpoint API functions take a (sometimes) optional query
string. Rather than pass in each of the query parameters into the API
functions separately, have them just take the query string (or NULL).

Passing each parameter in separately was OK with a couple of parameters,
but some take at least four which starts getting a bit unwieldy and
maybe there will be more...

This of course is an API break and as such the libraries minor version
has been bumped (as is the case for the 0.x devel series).

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>