curler: Handle re-directs
Some endpoints after a POST/PUT will re-direct (GET) to some other
endpoint if the request succeeds, e.g

    /self-assessment/ni/{nino}/{taxYear}/intent-to-crystallise

will currently (it's due to change) re-direct to

    /self-assessment/ni/{nino}/calculations/{calculationId}

Handle this by catching the 303 (See Other) response code and getting
the location url from curl and performing the request again.

We don't set CURLOPT_FOLLOWLOCATION, so we can do the re-direct
ourselves, this allows us to return both responses to the user and
putting the responses in a JSON array, e.g

    $ ./mtd-cli sa cr-intent-to-crystallise 2018-19
    [
        {
            "status_code": 303,
            "status_str": "See Other",
            "url": "https://test-api.service.hmrc.gov.uk/self-assessment/ni/XX1234567/2018-19/intent-to-crystallise",
            "method": "POST",
            "result": null
        },
        {
            "status_code": 200,
            "status_str": "OK",
            "url": "https://test-api.service.hmrc.gov.uk/self-assessment/ni/XX1234567/calculations/c2c82d00-c407-4f98-ab7d-c8319c522e6b",
            "method": "GET",
            "result": {
                ...
            }
        }
    ]

In fact all responses are now provided in an array even for single
responses. It was a toss up whether to only do arrays for multiple
responses, but decided to just be consistent and always use an array.

Technically, due to the change in JSON format, this constitutes an API
break.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>