mtd, auth: Fix oauth.json
With commit 34b21f6 ("curler: Handle re-directs") the JSON responses are
wrapped in an array including the oauth JSON, which meant that
refresh_access_token() couldn't get the refresh token from oauth.json.

We don't really need the oauth.json to be wrapped in an array so when
writing out this file extract the core JSON from the array.

Fixes: 34b21f6 ("curler: Handle re-directs")
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>