curler: Fix re-sending the data when retrying a request
If we are doing a POST/PUT with data from a file and we need to retry
the request because perhaps we had to refresh the access token, then
we'd fail to send any data the second time around as the file pointer
will now be at the end of the file.

Reset it in curl_perform().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>