curler: Store the location url in the result when required
Some endpoints return a 202 Accepted when they are going to do some
processing and the result won't be immediately available. In this case
they provide a "Location:" header (note as stated above, it's not a
re-direct) where to find the result and an estimated time when the
result will be ready to be viewed.

In these cases, catch that location (we store it in a new pointer in
struct curl_ctx so as not to interfere with the location pointer for
doing actual re-directs) and store it as a new field in the result JSON.
E.g

   $ mtd-cli sa tc-calculate ../test_files/tc-calculate.json
   [
       {
           "status_code": 202,
           "status_str": "Accepted",
           "url": "https://test-api.service.hmrc.gov.uk/self-assessment/ni/XX123456X/calculations",
           "method": "POST",
           "result": {
               "etaSeconds": 5,
               "location": "/self-assessment/ni/XX123456X/calculations/77427777"
           }
       }
   ]

So we can see that in an estimated 5 seconds the result will be
available, and we can see the (in this case) calculationId is '77427777'
and thus when the time is right we can execute

    $ mtd-cli sa tc-get-calculation 77427777

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>