Split the nino out into its own file
Move the NINO (National Insurance Number) out of config.json into its
own nino.json.

This is probably a good thing to do anyway as it's not clear yet exactly
how the client_id and client_secret will be handled.

Also this will aid in testing where if we start from a blank slate, we
won't have a NINO until we create a test person/business, when calling
mtd_init_config)_ we could put in a dummy NINO, but then there was no
way to change it, other than (well, editing the file) re-doing the
config again once we have it.

Now we can set it independently of anything else, though perhaps a
little more work until this is truly usable...

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>