mtd: Remove check_files()
This checked that the various config files existed whenever the library
was initialised. However that gets in the way due to some endpoints
needing to be called (albeit generally just for testing) when we don't
yet have an oauth.json or nino.json file.

Since commit 8444fe0 ("auth: Log errors when loading the JSON in
load_token()") we now log an error if a missing config file is attempted
to be accessed.

This removes an exported enum (MTD_ERR_MISSING_CONFIG) value so again is
an API break...

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>