Properly handle 'application' scope endpoints
While these endpoints work fine with a user token there are cases where
you need to call an application scope endpoint before you have any user
token, e.g when creating a test user.

Create another function analogous to refresh_access_token (which is
renamed to oauther_refresh_access_token), oauther_get_application_token
which will obtain an application scope OAuth token and store it in
oauth-application.json. These tokens don't have the refresh mechanism
like the user tokens do hence the need to a separate function for this.

We then make use of the scope member of struct _endpoint add in commit
6d79cd8 ("endpoints: Add a new member to struct _endpoint") to determine
which function should be used to obtain an access token and which JSON
file to find the token in.

We move enum oauth_scope out of endpoints.h into endpoints.c as that's
the only place it's used.

Finally we make a slight change in ep_make_url() to only lookup the NINO
if we need to (to avoid can't open file errors when we don't have
nino.json yet).

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>