logger: Always log __func__
Some of the log messages used __func__ to show the function that the log
message comes from.

Have this done for all log messages automatically.

This is accomplished by creating a macro wrapper for logger() that
passes __func__ into it. We need the macro for __func__ to show the
originating function of the message and not of logger() itself.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>