Makefile: Fix rpm build
'make rpm' was broken by commit 85e41ae ("src/Makefile: Get the full
version"), due to when building the rpm the source is no longer under
git and there's no .git directory so 'git describe' fails.

Fix this with a bit of trickery.

When doing 'make rpm' we put the output of 'git describe | tail -c +2'
into a .version file in the repository root, this is also gitignored as
it's only used for building the rpm.

We then append this file into the tar archive created by 'git archive
...' we need to use the --transform option to tar so we can add in the
tar prefix path as created by 'git archive ...'.

In src/Makefile we adjust our command for getting the version string to
try 'cat ../.version' if the 'git describe' fails (which it will when
building the rpm).

Fixes: 85e41ae ("src/Makefile: Get the full version")
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>