endpoints: Fix a crash in ep_make_url()
When adding a new endpoint I was seeing a segfault in ep_make_url() when
testing one of the newly added endpoint functions.

Turns out this was due to the params[] pointer being NULL, this pointer
is passed in as ctx->params, however this was NULL.

In do_ep() we build up the params array of pointers, however if we
didn't have any then we never set ctx.params to it and so ctx.params
remains NULL then when we try and access params[] members we get a
segfault.

This only shows up with endpoints that take an optional query string and
no other variables to substitute into the URL and we're not providing
any query string.

There are a few endpoints like this however they were added and tested
before the commit that caused this breakage.

So, long story short, we need to always set ctx.params in do_ep()
whether we have any params or not (in ep_make_url() we do check that the
individual params members are not NULL before trying to use them).

Fixes: 5ea027e ("endpoints: Add a generic do_ep() function")
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>