Create an internal mtd_ctx structure using Thread Local Storage
This is a continuation of the work started in commit af8e3e1 ("Prepare
for splitting out curl_global_init()") to allow for some global state to
be set in libmtdac while being thread safe.

This creates a new 'struct mcd_ctx' using thread local storage (TLS) to
replace the few global settings in mtd.c

We put a call to curl_global_init() in mtd_global_init(). Now either
mtd_global_init() or mtd_init() with the MTD_OPT_GLOBAL_INIT flag set
must be called and is thus an API break.

We set the log level and application connection type in this structure
now and its use will be expanded such as by adding the ability to set
custom request headers.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>