Allow to use a buffer for the JSON payload rather than a file
Currently when sending JSON data to the API endpoints this is done via a
file. Sometimes it may be more convenient to be able to use a buffer as
the source for the data instead.

This makes use of the mtd_ctx thread local storage infrastructure to
allow setting a pointer to a buffer and the size of the data in the
buffer to be sent.

Thus two new functions have been introduced

    mtd_set_src_data()

this takes a pointer to the buffer and the size (in bytes) of the data
in the buffer to send.

There is also

    mtd_unset_src_data()

that sets the buffer pointer and size to NULL/0 to allow for using files
again. Currently that is not strictly needed as if the src_file
parameter is not NULL then that will be used regardless if the buffer
pointer is set or not, but no guarantees that will always be the case.

This also happens to fix a minor issue of sending a Content-Type header
field when doing a GET after a POST/PUT.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>