Totally re-work how the source JSON payload is specified
This re-works how the JSON payload data for POST/PUT endpoints is
specified.

Originally this was simply done by passing a filename into the endpoint
API functions.

That was changed in commit 30765cb ("Allow to use a buffer for the JSON
payload rather than a file") to allow alternatively specifying a buffer
that contained the JSON, this was done using the mtd_ctx thread local
storage infrastructure.

This change basically undoes that and then goes much further.

Rather than passing a filename (or using the TLS buffer) into the
endpoint API functions. This has been replaced with a new mtd_dsrc_ctx
structure.

This can be used to specify a filename or a buffer, but also a stdio
FILE * file pointer or a file descriptor of already open files. Both of
which are useful mechanisms.

The mtd_dsrc_ctx structure holds a union of the above methods and then
you specify which method you are actually using in mtd_dsrc_ctx. This
new API looks like (defined in include/libmtdac/mtd.h)

    enum mtd_data_src_type {
        MTD_DATA_SRC_FILE = 0,
        MTD_DATA_SRC_BUF,
        MTD_DATA_SRC_FP,
        MTD_DATA_SRC_FD,
    };

    typedef union mtd_data_src {
        const void *buf;
        const char *file;
        FILE *fp;
        int fd;
    } mtd_data_src_t;

    struct mtd_dsrc_ctx {
        mtd_data_src_t data_src;
        size_t data_len;

        enum mtd_data_src_type src_type;
    };

The buf and file pointers are const as that is what you will often have
and thus avoids needing casting.

This will also allow for additional methods to be added, e.g it should
be possible to add the ability to specify a callback function to get the
data.

This is documented in the API.md and a new mtd_dsrc_ctx.3 man page.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>