curler: Flow the X-Correlationid header value through
Many (if not all) endpoint responses carry a X-Correlationid header,
which is described as a 'Unique id for operation tracking'.

Some applications may find this useful so provide it in the JSON output
under an 'id' field.

As part of the re-working of header_cb() we switch to using
strcasestr(3) for case insensitive searching as in the documentation the
header is called 'X-CorrelationId' whereas what is currently coming
through is 'X-Correlationid' (lower case i).

In set_response() when creating the JSON we make use of the '?' modifier
in json_pack() to either print the object or a null rather than having
to do a test...

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>