mtd: Allow the user to override the connection type
Previously we just set the application connection type to
MTD_ACT_OTHER_DIRECT.

This is still the default, but allow the user to override this by
passing one of the MTD_OPT_ACT_ flags into mtd_init().

Currently setting this to something else will cause no fraud prevention
headers to be sent, but this is the first step in allowing the user to
have fraud prevention headers sent for all connection types and to also
allow the user to provide their own values for the various headers if so
desired.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>