fph: Make *curl a thread local variable
Rather than define *curl in set_anti_fraud_hdrs() and then pass into the
various functions that need it, define it as a thread local variable.

This is simply required for the use of the curl_easy_escape() function.
Defining it as a global variables means we can make the get_*() function
only take a char *.

This is a part of the work to allow users to send fraud prevention
headers for application connection types other than MTD_ACT_OTHER_DIRECT
and also to provide their own function for generating the various
headers values.

The *curl thing is an internal implementation detail and should not be
leaked to the application, thus we will have a standard function
prototype for generating the various header values of

    char *get_<value>(char *buf);

We will have a structure containing function pointers to the various
get_*() functions that will default to internal implementations but will
be overrideable by the user with their own functions.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>