fph: Convert to a struct mtd_fph_ops of function pointers
This is part of the work to allow users to use connection types other
than OTHER_DIRECT and to allow them to be able to set their own fraud
prevention header values rather than the built in ones.

We create a new struct mtd_fph_ops (part of the user visible API, it
will be documented when the user can make use of it) that contains
function pointers for each fraud prevention header value setter
function.

These default to the current set of get_*() functions as used by the
OTHER_DIRECT connection type. Further work will allow the user to
override these functions with their own and expand on the set of get_*()
functions to cover the other connection types.

This has changed how these get_*() functions work slightly, rather than
take a buffer to be filled in as an argument, they simply return a
pointer to a dynamically allocated buffer. This will simplify the
interface for the users creating their own get_*() functions.

We create a new helper function (add_fph()) that wraps the getting of a
value, the setting of the header and the free'ing of this buffer.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>