fph: Allow users to override the fraud prevention headers
By default the fraud prevention headers are generated from internal
functions. However it may be useful to allow users to override these and
provide their own functions for generating the various header values.
There may be some headers which make more sense to come from the
application rather than the library.

To do this we change the prototype of mtd_init() to also take a new
struct mtd_cfg, currently, this structure just has a pointer to another
structure, a struct mtd_fph_ops, the user can define a new struct
mtd_fph_ops and set their own functions in it then link that to struct
mtd_cfg and pass that into mtd_init()

More members will be add to mtd_cfg, probably starting with the extra
headers that can currently be set via mtd_hdrs_add().

If mtd_cfg is passed in as NULL then all defaults will be used, likewise
if mtd_cfg.fph_ops is NULL or any of its members, then the default fraud
prevention header functions will be used.

While we're at it we also change the flags argument from int to unsigned
int, this seems preferable for flags arguments and also gives an extra
bit for a flag.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>