mtd: Re-work struct mtd_ctx initialisation
In order to make something like

    mtd_init()
    ...
    mtd_deinit();
    mtd_init();
    ...

in the same thread to work properly we want to initialise struct mtd_ctx
in mtd_init() rather than just at thread creation time.

Create a struct dfl_mtd_ctx with default values and use this to
initialise struct mtd_ctx in mtd_init();

Do some cleanup of the setting of the connection type seeing as we set a
default value for it.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>