mtd/mtd_init_auth: Factor out the api scopes
Rather then specify the various scopes required by the library directly
when setting the url, split them out into an array of strings.

This just makes it a little clearer at first glance what is being
requested and as more are added saves the url string in the code from
growing.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>