endpoints: Re-work endpoints[]
This is an array of struct _endpoints's that describe each of the
supported API endpoints.

The first member of this structure was an enum saying which endpoint it
is. This member is not actually used within the code.

Instead, use the endpoint enums as the array indexes, this provides both
the _documentation_ that the .ep member did as well as being more
explicit about the direct connection between endpoints[] and enum
endpoint.

It also removes a good number of lines of code while providing the same
thing.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>