logger: Give logger perror() like capability
perror(3) is a handy function for returning the descriptive text for
various errno values.

However it only prints to stderr and has limited message formatting.
There are times when I'd like to use perror() but tie it in with the
libmtdac logging.

So this is what this does, we add a new MTD_LOG_ERRNO log level which
triggers the new behaviour.

This makes use of strerror_r(3) to do the errno -> text translation and
nicely incorporates it into our existing logging, e.g

    [libmtdac ERROR] try_connect: No route to host

that can be produced with

    logger(MTD_LOG_ERRNO, NULL);

and it picks up the errno value. You can also pass the normal string
format in instead of NULL to get more tailored messages. e.g

    logger(MTD_LOG_ERRNO, "more context here:");

which would give

    [libmtdac ERROR] more context here: try_connect: No route to host

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>