curler: Have libcurl use our own socket
libcurl has the ability to be passed an already connected socket, rather
than doing it itself internally.

We need this ability so we can provide the Gov-Client-Public-IP &
Gov-Client-Public-Port fraud prevention headers.

These specify the source IP address and port of the connected socket and
we need to know these before we invoke libcurl.

This just makes libcurl use our own connected socket the, further
commits will deal with the above headers.

One change in do_curl() is that if we are re-trying the connection we
always re-calculate the headers as this will use a new connection and at
the least the above header values will change.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>