valgrind: Add a suppressions file for getaddrinfo
With glibc 2.32 on Fedora 33 valgrind is producing numerous 'still
reachable' warnings, this seems to be caused by the call to
getaddrinfo(3).

Just have them suppressed for now. This is done via a suppressions file
(src/.valgrind-getaddrinfo-supp) that can be passed into valgrind via
--suppressions=

This was generated by

    valgrind --leak-check=full --show-leak-kinds=all --gen-suppressions=yes

and then trimmed down.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>