Add and use our own percent encoding function
For the fraud prevention headers various values need to be percent
encoded.

Previously we were using curl_easy_escape() todo this, worked fine, but
it did mean we needed quite a bit of curl stuff in here just to use that
function as it requited passing in a CURL context.

Add out function for doing this and switch over to it so we can actually
remove all the curl stuff, which is otherwise unnecessary, from fph.c.

This new function is exported as it will be useful for clients to use
when we have the ability for them to set a _client_ version string to go
along with the libmtdac version string that is sent in the
'Gov-Vendor-Version' header.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>