fph: Add the ability to set a client version string
There is already fph_version() for setting a version string pertaining
to libmtdac. However it would be useful if there was a way for a client
to also set a version string.

This can be accommodated for as the spec allows for the following format

    <software-name>=<version-number>&<software-name-2>=<version-number-2>& ...

So, if a client leaves mtd_fph_ops.fph_version unset but sets
mtd_fph_ops.fph_version_cli to a function that returns a string in the
form

    client_name=client_version

where client_name & client_version are percent encoded,
(mtd_percent_encode() can be used for this), this will be appended to
the libmtdac version string in the above format.

You can of course just define mtd_fph_ops.fph_version and then set the
version from there without needing .fph_version_cli.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>