fph: Add a helper macro for setting the fph_ops functions
This macro can be used for setting the function in struct mtd_fph_ops,
it can be used with the #define'd member names which some people may
find more convenient to use.

It takes a struct mtd_sph_ops, the name of the structure member to set
and the function to set it to.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>