curler: Fix formatting of real numbers in the JSON response
When getting JSON replies containing real numbers we were getting them
displayed like

    5021.1099999999997

instead of

    5021.11

this is of course due to the fact that real numbers can't always be
expressed exactly in binary.

We can get the desired result by adjusting the precision, 16 seems to
mostly do the right thing, 15 does better but 15 is not enough to allow
for all the number ranges required, i.e it won't accurately represent
99999999999999.98 rounding it up to the next whole number.

We can tell jansson to use 16 digits of precision for real numbers in
set_response().

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>