curler: Fix PUT with data from a buffer
The ability to have the POST/PUT data come from a buffer rather than a
file was supposedly added by commit 30765cb ("Allow to use a buffer for
the JSON payload rather than a file"), I say _supposedly_, as looking at
the code I'm not sure it could have worked for PUT. Then commit 08eb2cf
("Totally re-work how the source JSON payload is specified")
significantly re-worked that, but again I don't see how it could have
worked... POST would have been OK.

One of the issues is that if we had the data in a buffer and were doing
a PUT then nothing told curl that we were doing a PUT, if we had it in
a file we would set CURLOPT_UPLOAD to 1.

Re-work some of the code in curl_perform() to fix this. Nice bit of code
reduction/simplification to boot!

Rather than using CURLOPT_UPLOAD to inform curl we want to do a PUT,
explicitly set it as a PUT request with CURLOPT_CUSTOMREQUEST.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>