endpoints: Set buf to NULL earlier in do_ep()
buf is used to pass the resultant JSON from the curl API call back to
the user.

We make no statement about whether the user should initialise this to
NULL first, rather we do it in do_ep().

However we could error out before we initialised it and the user would
then perhaps try to free an invalid pointer.

We should initialise buf to NULL as the first thing we do. free(NULL) is
perfectly valid.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>