mtd: Add a function to get the HTTP status code of a request
I came across a need to check the HTTP status code of a request after
getting a MTD_ERR_REQUEST to see if it was a 404.

This function takes a JSON result buffer and returns the status code
(which can be checked against the MTD_HTTP_* enum values) of the last
request contained in the buffer as this is likely to be the one that you
are interested in as in the case of MTD_ERR_REQUEST you may need to
check what actually caused the request to fail.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>