mtd: Re-work casting in mtd_percent_encode()
mtd_percent_encode() allows to pass the string length as -1 to indicate
to use strlen(3) to get the length. We pass that in as a ssize_t.
However we use size_t for the actual percent encoded buffer length, that
is what strlen returns and malloc(3) takes.

If we don't have any casting in the if?: operation (to work out if we
use the passed in length or strlen) we get

mtd.c: In function ‘mtd_percent_encode’:
mtd.c:175:37: warning: operand of ‘?:’ changes signedness from ‘ssize_t’ {aka ‘long int’} to ‘size_t’ {aka ‘long unsigned int’} due to unsignedness of other operand [-Wsign-compare]
  175 |  buflen = ((len < 0 ? strlen(str) : len) * 3) + 1;
      |                                     ^~~

So we cast the strlen to ssize_t, that works, no casting would be better
but lets turn the if?: around so we cast the passed in length instead,
at that point we know we are using it for the length and it won't be
-ve. That way we're not casting strlen away from what it is.

The other option would be to just pass the length in as a size_t and use
0 to indicate to use strlen, but for me, -1 seems better for this
indication.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>