fph: Allow to pass user data to the various fph_* functions
It may be useful to be able to pass some data into the fph_* functions
for setting the various fraud prevention headers.

This can now be done by setting the user_data pointer in struct
mtd_fph_ops, this will then be passed into any fph_ functions you set.

It is ignored by most of the internal functions.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>