logger: Properly handle errors from vasprintf(3)
Don't check the status of logbuf, we init it to NULL and after an error
from vasprintf() its state will be undefined.

Given that, just check for a return of -1 and if so, free logbuf which
may or may not have been allocated (in which case it will still be NULL)
and return.

We can also do the len check _after_ we call va_end(3) for further
simplification.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>