Contributing.md: Re-work to incorporate the DCO
The Developer's Certificate of Origin was introduced by the Linux kernel
and is where the 'Signed-off-by' stuff came from.

It's a _not_ a CLA, but just asks the developer to confirm they have the
right to submit their code to the repository.

Add a copy of the DCO to the repository and explain about it in
Contributing.md.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>