mtd: Have applications specify the required scopes
When an application authorises with HMRC it needs to specify the OAuth
2.0 'scopes' required.

Previously libmtdac would list all the scopes that it supports. However
it turns out this is not desirable (even though it's not mentioned in
the documentation and even the example shown lists more than two
scopes). This was pointed out in the initial demonstration to HMRC.

This commit changes things so that now applications will tell libmtdac
which scopes they require, e.g for Income Tax Self-Assessment;
read:self-assessment&write:self-assessment

This is done by a new parameter to mtd_init_auth() which now takes a
bitwise OR of the required scopes.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>