mtd: Introduce per API authorisation scope namespace
During an initial demo with HMRC it came to light that libmtdac was
requesting too many OAuth scopes.

Now AFAICS there isn't anything in the documentation that says you
shouldn't request for example itsa & vat scopes, but apparently this is
bad.

The first part of the solution for this was done in commit e9795bd
("mtd: Have applications specify the required scopes").

That works for itsa[0] which only needs read/write access to the
self-assessment scope, so that's a good thing that it now only has
access to the bits it needs. However mtd-cli[1] being a wrapper around
libmtdac needs access to self-assessment, vat and national-insurance
scopes (the latter not technically being a part of MTD and may be
removed from libmtdac at some point)..

The idea I've come up with to deal with needing access to multiple
different scopes (hopefully this will pass muster) is to allow to have
multiple sets of authorisation tokens in oauth.json, one set for each
API you wish to use, so itsa and vat would have their own sets of access
tokens.

This seems to work OK, you can go through the OAuth authorisation
process multiple times using the same login details but requesting
different scopes.

So now the oauth.json can look like

  {
      "itsa": {
          "access_token": "60b725f10c9c85c70d97880dfe8191b3",
          "refresh_token": "3b5d5c3712955042212316173ccf37be",
          "expires_in": 14400,
          "scope": "write:self-assessment read:self-assessment",
          "token_type": "bearer"
      },
      "vat": {
          "access_token": "2cd6ee2c70b0bde53fbe6cac3c8b8bb1",
          "refresh_token": "e29311f6f1bf1af907f9ef9f44b8328b",
          "expires_in": 14400,
          "scope": "write:vat read:vat",
          "token_type": "bearer"
      }
  }

The main user visible change is that mtd_init_auth() now looks like

  int mtd_init_auth(enum mtd_ep_api api, unsigned long scopes);

The first argument specifies the API to add scopes for; itsa, vat etc.
The second argument specifies what scopes to authorise.

[0]: https://github.com/ac000/itsa
[1]: https://github.com/ac000/mtd-cli

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>