auth: Convert to write_config()
There is a function in mtd.c called write_config which is a wrapper
around json_dumpfd(). It uses a combination of open(2) and openat(2) to
open a fd for the required file to write to. It's used by a few
functions in mtd.c.

Lets also use it in auth.c for consistency.

Closes: https://github.com/ac000/libmtdac/issues/5
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>