mtd: Introduce per API credentials
In the process of trying[0] to acquire production credentials for
mtd-cli[1] & itsa[2] it came up in discussions that getting credentials
for mtd-cli might be slightly more tricky due to it supporting both ITSA
& VAT APIs.

Perhaps however we can ease this slightly by splitting mtd-cli into two
_virtual_ parts for each of the two APIs, each with their own set of
credentials.

This commit extends mtd_init_creds() & load_token() to store and lookup
credentials under a specific API. So instead of the creds.json looking
like

  {
      "client_id": "YWJjZGVmZ2hpamtsbW5vcHFyc3QK",
      "client_secret": "0773a244-e434-4dac-a579-8134c5691754"
  }

it will now look like (with credentials for two APIs)

  {
      "itsa": {
          "client_id": "YWJjZGVmZ2hpamtsbW5vcHFyc3QK",
          "client_secret": "0773a244-e434-4dac-a579-8134c5691754"
      },
      "vat": {
          "client_id": "cXdlcnR5dWlvcGFzZGZnaGprbHoK",
          "client_secret": "9479b310-6775-468c-b918-a9512c4f7622"
      }
  }

This adds a parameter to the mtd_init_creds() function that specifies
what API to add credentials for.

[0]: https://github.com/ac000/libmtdac/discussions/18
[1]: https://github.com/ac000/mtd-cli
[2]: https://github.com/ac000/itsa

Closes: https://github.com/ac000/libmtdac/issues/28
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>