Factor out Linux/Glibc specific stuff
As part of the port to FreeBSD, factor out the Linux specific stuff into
a platform/linux directory, the Makefile will select the right platform
directory at build time.

This comprises the code to

  - generate a UUID (platform/linux/gen_uuid.c)
  - extract a link-level (MAC) address from struct ifaddrs
    (platform/linux/get_ll_addr.c)
  - get the machine vendor and model for the Gov-Client-User-Agent
    header (platform/linux/get_mach_info.c)

We also need to handle strerror_r(3), on Linux/Glibc there are two
versions of this function. The GNU version which returns a pointer to
the result buffer and a POSIX (XSI) version which returns an int error
code.

FreeBSD only has the XSI version. To get the XSI version on Linux you
need to _not_ specify _GNU_SOURCE, however we make extensive use of
that.

Also the GNU semantics are quite useful and we make use of the fact it
returns a pointer to the result buffer.

To solve this we create a simple wrapper around the XSI version of
strerror_r which is in its own file (platform/common/x_strerror_r.c)
which avoids the definition of _GNU_SOURCE, which provides the GNU
semantics and will work on both Linux & FreeBSD, hence the 'common'
directory.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>