src/Makefile: Re-work setting of linker flags
When compiling under older llvm (8.0.1) on FreeBSD 12.1 we got a linker
error

  ld: error: unknown -z value: undefs

When linking with '-z defs' under FreeBSD 13 with llvm 11.0.1 we had
previously got a linker error about 'undefined symbol: environ'. So when
building under FreeBSD we replaces '-z defs' with '-z undefs' which
worked,

That causes an issue with older llvm. However we can simply remove that
option altogether. So on Linux we set it (which AIUI is a good thing to
do), but not on FreeBSD.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>