Setup Travis CI
This adds an initial .travis.yml file that does three builds.

We do two builds with GCC on Linux and one build with Clang on FreeBSD.
All x86-64.

For GCC we use versions 4.8.2 & 11. These are done under pretty old
Ubuntu.

We use 4.8.2 as that's close to 4.8.5 that's on CentOS 7 which is the
sort of oldest Linux I try and support.

11 as that's the current version, even though I'm generally building
with a current'ish GCC on Fedora.

Clang (currently 8.0.1) on FreeBSD to get that different compiler/OS
coverage.

We compile with V=1 passed to make to get the full compiler/linker
commands, useful if you need to debug build issues.

One other thing to note is that by default travis will only clone the
last 50 commits of the repository, this causes an issue here as part of
the build process does a 'git describe --long --dirty' to get the
version, but because there are no tags within the last 50 commits, git
can't find a name to use so the version ends up blank. This doesn't
break the build, but does result in compiler warnings due to doing
'GIT_VERSION + 1'. So we just do a full clone of the repository.

The Travis CI dashboard for libmtdac can be found here,
https://app.travis-ci.com/github/ac000/libmtdac

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>