Enable linking with -zdefs on FreeBSD
We couldn't link with -zdefs under FreeBSD due to the environ variable
(which we use in libmtdac) which unlike in Linux/glibc is _not_ defined
in libc but in /usr/lib/crt1.o which is only linked into executables.

That means linking libmtdac always produces

  ld: warning: undefined symbol: environ

However -zdefs is really useful for finding legitimate issues.

So what we do on FreeBSD is link with -zdefs as we do on Linux, but turn
any undefined symbols errors into warnings so the build doesn't fail.

For CI environments we do want the build to fail for any undefined
symbols that aren't 'environ'.

To enable that we create a script to wrap the build process on FreeBSD.
This script looks for undefined symbols, if we get no undefined symbols
or only environ, return 0 (success) if we get any other undefined
symbols return 1 (failure).

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>